feat(tablepanel): refactor TablePanel component
