+++ src/app/components/TablePanel/Loadable.ts
@@ -10,13 +10,3 @@ export const TablePanel = lazyLoad(
   () => import('./index'),
   module => module.default,
 );
-
-export const TabLabel = lazyLoad(
-  () => import('./Label'),
-  module => module.TabLabel,
-);
-
-export const TipLabel = lazyLoad(
-  () => import('./Label'),
-  module => module.TipLabel,
-);
