+++ src/app/components/TablePanel/index.tsx
@@ -1,7 +1,220 @@
-import Panel from './Panel';
-import { ColumnsType } from './Panel';
-import { TabLabel, TipLabel } from './Label';
+import React, { useEffect } from 'react';
+import { useTranslation } from 'react-i18next';
+import { translations } from '../../../locales/i18n';
+import { Table, Pagination, Loading } from '@cfxjs/react-ui';
+import { useHistory, useLocation, useParams } from 'react-router-dom';
+import queryString from 'query-string';
+import PropTypes from 'prop-types';
+import useSWR from 'swr';
+import { simpleGetFetcher } from '../../../utils/api';
+import styled from 'styled-components';
+import { media } from '../../../styles/media';
+import { PaginationProps } from '@cfxjs/react-ui/dist/pagination/pagination';
+import { Props as TableProps } from '@cfxjs/react-ui/dist/table/table';
+import { useBreakpoint } from '../../../styles/media';
 
-export default Panel;
-export type { ColumnsType };
-export { TabLabel, TipLabel };
+const StyledPaginationWrapper = styled.div`
+  margin: 1.7143rem 0;
+  li button {
+    background-color: rgba(0, 84, 254, 0.04);
+  }
+  .input-wrapper.solid,
+  div.select {
+    background-color: rgba(0, 84, 254, 0.04) !important;
+    border-color: transparent;
+    &.hover,
+    &:hover {
+      background-color: #e0eaff !important;
+      border-color: transparent;
+    }
+    input {
+      color: #74798c !important;
+      font-size: 1rem;
+      font-family: CircularStd-Medium, CircularStd;
+      font-weight: 500;
+    }
+  }
+
+  div.text,
+  button,
+  div.option span {
+    font-size: 1rem !important;
+    font-family: CircularStd-Medium, CircularStd;
+    font-weight: 500;
+    color: #74798c !important;
+    line-height: 1.2857rem !important;
+  }
+  button.active {
+    color: #fff !important;
+  }
+`;
+
+const StyledTableWrapper = styled.div`
+  .table-content {
+    padding: 0 1.2857rem 1.2857rem;
+  }
+  .table th.table-cell {
+    white-space: nowrap;
+  }
+  .table td.table-cell {
+    font-size: 1rem;
+    font-family: CircularStd-Book, CircularStd;
+    font-weight: 400;
+    color: #20253a;
+    padding: 1.2857rem calc((0.5714rem / 2) * 3);
+    white-space: nowrap;
+    line-height: 1;
+
+    ${media.s} {
+      padding: 1.1429rem;
+    }
+  }
+`;
+
+type TablePanelConfigType = {
+  url: string;
+  pagination?: PaginationProps & boolean;
+  table: TableProps<unknown>;
+  onDataChange?: Function;
+};
+
+// pagination default config
+const defaultPaginationConfig: PaginationProps = {
+  total: 0,
+  page: 1,
+  pageSize: 10,
+  showPageSizeChanger: true,
+  showQuickJumper: true,
+  size: 'small',
+  variant: 'solid',
+  onPageChange: () => {},
+  onPageSizeChange: () => {},
+};
+// mobile pagination default config
+const defaultPaginationMobileConfig: PaginationProps = {
+  ...defaultPaginationConfig,
+  labelPageSizeBefore: '',
+  labelPageSizeAfter: '',
+  limit: 3,
+};
+// table default config
+const defaultTableConfig: TableProps<unknown> = {
+  data: [],
+  rowKey: 'key',
+  columns: [],
+  variant: 'solid',
+};
+
+function TablePanel({
+  url,
+  pagination,
+  table,
+  onDataChange,
+}: TablePanelConfigType) {
+  const { t } = useTranslation();
+  const breakpoint = useBreakpoint();
+  const history = useHistory();
+  const location = useLocation();
+  // merged pagination config
+  const paginationObject = typeof pagination === 'boolean' ? {} : pagination;
+  const mergedPaginationConfig = {
+    labelPageSizeBefore: t(translations.general.pagination.labelPageSizeBefore),
+    labelPageSizeAfter: t(translations.general.pagination.labelPageSizeAfter),
+    labelJumperBefore: t(translations.general.pagination.labelJumperBefore),
+    labelJumperAfter: t(translations.general.pagination.labelJumperAfter),
+    ...(breakpoint === 's'
+      ? defaultPaginationMobileConfig
+      : defaultPaginationConfig),
+    ...paginationObject,
+  };
+  const parsedUrl = queryString.parseUrl(url);
+  const query = {
+    page: mergedPaginationConfig.page,
+    pageSize: mergedPaginationConfig.pageSize,
+    ...parsedUrl.query,
+    ...queryString.parse(location.search),
+  };
+
+  const { data, error } = useSWR([url + location.search], simpleGetFetcher);
+
+  useEffect(() => {
+    onDataChange && onDataChange({ data, error });
+  }, [data]); // eslint-disable-line
+
+  let emptyText: React.ReactNode | string = t(
+    translations.general.table.noData,
+  );
+  let tableData = table.data;
+  let paginationTotal: number = 0;
+
+  if (!data && !error) {
+    emptyText = <Loading />;
+  }
+
+  if (data && !error) {
+    tableData = data.result?.list || table.data;
+    paginationTotal = data.result?.total || defaultPaginationConfig.total;
+  }
+
+  const handlePaginationChange = (page: number, pageSize: number): void => {
+    const search = queryString.stringify({
+      ...parsedUrl.query,
+      ...queryString.parse(location.search),
+      page,
+      pageSize,
+    });
+    history.push(`${location.pathname}?${search}`);
+  };
+
+  return (
+    <>
+      <StyledTableWrapper>
+        <Table
+          tableLayout="fixed"
+          columns={table.columns}
+          rowKey={table.rowKey}
+          data={tableData}
+          emptyText={emptyText}
+          scroll={{ x: 800 }}
+        />
+      </StyledTableWrapper>
+      <StyledPaginationWrapper>
+        {pagination !== false && (
+          <Pagination
+            {...mergedPaginationConfig}
+            onPageChange={handlePaginationChange}
+            onPageSizeChange={handlePaginationChange}
+            page={Number(query.page)}
+            pageSize={Number(query.pageSize)}
+            total={paginationTotal}
+          />
+        )}
+      </StyledPaginationWrapper>
+    </>
+  );
+}
+
+TablePanel.defaultProps = {
+  url: '',
+  pagination: defaultPaginationConfig,
+  table: defaultTableConfig,
+  onDataChange: () => {},
+};
+
+TablePanel.propTypes = {
+  url: PropTypes.string,
+  pagination: PropTypes.shape({
+    page: PropTypes.number,
+    pageSize: PropTypes.number,
+    onPageChange: PropTypes.func,
+    onPageSizeChange: PropTypes.func,
+  }),
+  table: PropTypes.shape({
+    data: PropTypes.array,
+    columns: PropTypes.array,
+    rowKey: PropTypes.oneOfType([PropTypes.string, PropTypes.func]),
+  }),
+  onDataChange: PropTypes.func,
+};
+
+export default TablePanel;
